Split test_credit_tracking into focused modules under the 500 LOC limit
The single 986-LOC module exceeded the unit-test file-size guideline. Split it
by concern into self-contained modules (verbatim test bodies, identical
collected set):

- test_credit_tracking_sink.py        CreditSink running-minimum semantics
- test_credit_tracking_capture.py     x-credits-remaining capture via HTTP helpers
- test_credit_tracking_decorator.py   @pro_api_credit_scope decorator
- test_credit_tracking_note.py        build_tool_response low-credits note
- test_credit_tracking_end_to_end.py  end-to-end MCP-mode note flow

Update the project tree in AGENTS.md accordingly.
